+++ mlir/test/Conversion/MIGraphXToTosa/mixr-to-tosa-ops.mlir
@@ -525,7 +525,7 @@ module {
   // CHECK-DAG: [[RECIPROCAL:%[a-z0-9]+]] = "tosa.reciprocal"([[REDUCE_SUM]])
   // CHECK-DAG: "tosa.mul"([[EXP]], [[RECIPROCAL]])
   func.func @func_softmax_1d(%arg0: !migraphx.shaped<16xf32, 1>) -> !migraphx.shaped<16xf32, 1> {
-    %0 = migraphx.softmax %arg0 {axis = 0 : i64} : <16xf32, 1> -> <16xf32, 1>
+    %0 = migraphx.softmax %arg0 {axis = 0 : i32} : <16xf32, 1> -> <16xf32, 1>
      return %0 : !migraphx.shaped<16xf32, 1>
   }
 
@@ -537,7 +537,7 @@ module {
   // CHECK-DAG: [[RECIPROCAL:%[a-z0-9]+]] = "tosa.reciprocal"([[REDUCE_SUM]])
   // CHECK-DAG: "tosa.mul"([[EXP]], [[RECIPROCAL]])
   func.func @func_softmax_4d(%arg0: !migraphx.shaped<16x16x16x16xf32, 4096x256x16x1>) -> !migraphx.shaped<16x16x16x16xf32, 4096x256x16x1> {
-    %0 = migraphx.softmax %arg0 {axis = 1 : i64} : <16x16x16x16xf32, 4096x256x16x1> -> <16x16x16x16xf32, 4096x256x16x1>
+    %0 = migraphx.softmax %arg0 {axis = 1 : i32} : <16x16x16x16xf32, 4096x256x16x1> -> <16x16x16x16xf32, 4096x256x16x1>
      return %0 : !migraphx.shaped<16x16x16x16xf32, 4096x256x16x1>
   }
 }
