+++ mlir/test/fusion/pr-e2e/attention/mixr-attention-padded-scale.mlir
@@ -5,7 +5,7 @@ module {
   func.func private @mlir_attention(%arg0: !migraphx.shaped<1x7x3xf32, 21x3x1> {func.read_access}, %arg1: !migraphx.shaped<1x3x7xf32, 21x7x1> {func.read_access}, %arg2: !migraphx.shaped<1x7x3xf32, 21x3x1> {func.read_access}, %arg3: !migraphx.shaped<1x7x7xf32, 49x7x1> {func.read_access}) -> (!migraphx.shaped<1x7x3xf32, 21x3x1> {func.write_access}) {
     %0 = migraphx.dot %arg0, %arg1: <1x7x3xf32, 21x3x1>, <1x3x7xf32, 21x7x1> -> <1x7x7xf32, 49x7x1>
     %scaled = migraphx.mul %0, %arg3 : <1x7x7xf32, 49x7x1>, <1x7x7xf32, 49x7x1> -> <1x7x7xf32, 49x7x1>
-    %1 = migraphx.softmax %scaled{axis = 2 : i64} : <1x7x7xf32, 49x7x1> -> <1x7x7xf32, 49x7x1>
+    %1 = migraphx.softmax %scaled{axis = 2 : i32} : <1x7x7xf32, 49x7x1> -> <1x7x7xf32, 49x7x1>
     %2 = migraphx.dot %1, %arg2: <1x7x7xf32, 49x7x1>, <1x7x3xf32, 21x3x1> -> <1x7x3xf32, 21x3x1>
     return %2 : !migraphx.shaped<1x7x3xf32, 21x3x1>
   }
@@ -18,7 +18,7 @@ module {
     func.func private @mlir_attention(%arg0: !migraphx.shaped<1x7x3xf32, 21x3x1> {func.read_access}, %arg1: !migraphx.shaped<1x3x7xf32, 21x7x1> {func.read_access}, %arg2: !migraphx.shaped<1x7x3xf32, 21x3x1> {func.read_access}, %arg3: !migraphx.shaped<1x7x7xf32, 49x7x1> {func.read_access}) -> (!migraphx.shaped<1x7x3xf32, 21x3x1> {func.write_access}) attributes {kernel, original_func = @mlir_attention} {
       %0 = migraphx.dot %arg0, %arg1: <1x7x3xf32, 21x3x1>, <1x3x7xf32, 21x7x1> -> <1x7x7xf32, 49x7x1>
       %scaled = migraphx.mul %0, %arg3 : <1x7x7xf32, 49x7x1>, <1x7x7xf32, 49x7x1> -> <1x7x7xf32, 49x7x1>
-      %1 = migraphx.softmax %scaled{axis = 2 : i64} : <1x7x7xf32, 49x7x1> -> <1x7x7xf32, 49x7x1>
+      %1 = migraphx.softmax %scaled{axis = 2 : i32} : <1x7x7xf32, 49x7x1> -> <1x7x7xf32, 49x7x1>
       %2 = migraphx.dot %1, %arg2: <1x7x7xf32, 49x7x1>, <1x7x3xf32, 21x3x1> -> <1x7x3xf32, 21x3x1>
       return %2 : !migraphx.shaped<1x7x3xf32, 21x3x1>
     }
