+++ mlir/test/fusion/pr-e2e/attention/mixr-attention.mlir
@@ -4,7 +4,7 @@
 module {
   func.func private @mlir_attention(%arg0: !migraphx.shaped<1x64x64xf32, 4096x64x1> {func.read_access}, %arg1: !migraphx.shaped<1x64x64xf32, 4096x64x1> {func.read_access}, %arg2: !migraphx.shaped<1x64x64xf32, 4096x64x1> {func.read_access}) -> (!migraphx.shaped<1x64x64xf32, 4096x64x1> {func.write_access}) {
     %0 = migraphx.dot %arg0, %arg1: !migraphx.shaped<1x64x64xf32, 4096x64x1>, !migraphx.shaped<1x64x64xf32, 4096x64x1> -> !migraphx.shaped<1x64x64xf32, 4096x64x1>
-    %1 = migraphx.softmax %0{axis = 2 : i64} : !migraphx.shaped<1x64x64xf32, 4096x64x1> -> !migraphx.shaped<1x64x64xf32, 4096x64x1>
+    %1 = migraphx.softmax %0{axis = 2 : i32} : !migraphx.shaped<1x64x64xf32, 4096x64x1> -> !migraphx.shaped<1x64x64xf32, 4096x64x1>
     %2 = migraphx.dot %1, %arg2: !migraphx.shaped<1x64x64xf32, 4096x64x1>, !migraphx.shaped<1x64x64xf32, 4096x64x1> -> !migraphx.shaped<1x64x64xf32, 4096x64x1>
     return %2 : !migraphx.shaped<1x64x64xf32, 4096x64x1>
   }
@@ -16,7 +16,7 @@ module {
   module @__xmodule_ attributes {mhal.arch = "##TOKEN_ARCH##", mhal.module} {
     func.func private @mlir_attention(%arg0: !migraphx.shaped<1x64x64xf32, 4096x64x1> {func.read_access}, %arg1: !migraphx.shaped<1x64x64xf32, 4096x64x1> {func.read_access}, %arg2: !migraphx.shaped<1x64x64xf32, 4096x64x1> {func.read_access}) -> (!migraphx.shaped<1x64x64xf32, 4096x64x1> {func.write_access}) attributes {kernel, original_func = @mlir_attention} {
       %0 = migraphx.dot %arg0, %arg1: !migraphx.shaped<1x64x64xf32, 4096x64x1>, !migraphx.shaped<1x64x64xf32, 4096x64x1> -> !migraphx.shaped<1x64x64xf32, 4096x64x1>
-      %1 = migraphx.softmax %0{axis = 2 : i64} : !migraphx.shaped<1x64x64xf32, 4096x64x1> -> !migraphx.shaped<1x64x64xf32, 4096x64x1>
+      %1 = migraphx.softmax %0{axis = 2 : i32} : !migraphx.shaped<1x64x64xf32, 4096x64x1> -> !migraphx.shaped<1x64x64xf32, 4096x64x1>
       %2 = migraphx.dot %1, %arg2: !migraphx.shaped<1x64x64xf32, 4096x64x1>, !migraphx.shaped<1x64x64xf32, 4096x64x1> -> !migraphx.shaped<1x64x64xf32, 4096x64x1>
       return %2 : !migraphx.shaped<1x64x64xf32, 4096x64x1>
     }
